When a product is clicked it will not longer pass the product informations like price in the query params. Instead just the productID and quantity is passed as query params.
Next the productID is used to call the backend and create a subscription for the checkout. After creation the product details are called with the subscription id.
The product details will call the a new api endpoint and fetch product informations like price and name etc.
So when a customer displays the subscriptions and one changes a price before the customer would not get shown the exact price. Now the price is fetched from the actual subscripton so it will always be accurate.