Completly redone auth service.
On email verify path the button to proceed to checkout first reloads the user and then checks if email was verified. While waiting for promise it buttons is deactivated and spinner is shown in button.
Moved all getHttpOptions() to authService.
In auth service user is not any longer any but from type User.