When opening change payment it will set a return url so after cancel or succeess the return url will be opened.
Added a cancel button to change payment.payment-changed-processing.
When after clicking on checkout the cancel action is called it returns to checkout and then redirected to change payment again.
Change payment status for success, processing and failed it will be able to open return url. When failed is called its only possible to go back to change payment.